feat: refactor booking package management and add new categories
- Updated booking validation to use packages-registry instead of studio-packages.
- Modified categories to include new categories: anniversaire and grad-series, and updated availability.
- Enhanced slot generation logic to accommodate new package durations.
- Introduced new service catalog entries for anniversaire and grad-series with corresponding images.
- Updated i18n dictionaries to include translations for new categories and services.
- Created new package files for anniversaire and grad-series, consolidating package management in packages-registry.
- Adjusted booking modal components to reflect changes in package sourcing and category handling.

+++ src/application/booking/booking-validation.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { contactSchema } from "./contact-validation";
-import { findPackageById, allowedDurations } from "./studio-packages";
+import { findPackageById, allowedDurations } from "./packages-registry";
 import { findCategory } from "./categories";
 
 const DATE_RE = /^\d{4}-\d{2}-\d{2}$/;
@@ -15,7 +15,7 @@ export const bookingSubmissionSchema = z
   .object({
     category: z.string().min(1),
     packageId: z.string().min(1),
-    durationHours: z.number().int().positive(),
+    durationHours: z.number().positive(),
     date: z.string().regex(DATE_RE, "invalid_date"),
     time: z.string().regex(TIME_RE, "invalid_time"),
     contact: contactSchema,
